scripts: avoid type shell builtin.
For maximum portability it might be wise to avoid shell builtins when testing
for a program's existence.

Reasoning:

`type` is considered an XSI extension[1] despite it being present in SVR2.

`command -v` is specified by posix but only on "systems supporting the User
Portability Utilities option"[2].

`type` is much older than `command` and probably better (`command` doesn't even
appear in SVR4, so who knows when it was created), but maybe we can just avoid
these altogether.

Curiously, the `hash` builtin is not marked as optional at all[3] and also
appears in the SVR2 man pages and source code.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/utilities/type.html
[2] https://pubs.opengroup.org/onlinepubs/009604599/utilities/command.html
[3] https://pubs.opengroup.org/onlinepubs/9699919799/utilities/hash.html